feat: small rework for GPO
