+++ ad_miner/sources/modules/objects.py
@@ -1,4 +1,7 @@
 import time
+import json
+
+from os.path import sep
 
 from ad_miner.sources.modules import logger
 from ad_miner.sources.modules.graph_class import Graph
@@ -12,9 +15,12 @@ from urllib.parse import quote
 
 
 class Objects:
-    def __init__(self, arguments, neo4j):
+    def __init__(self, arguments, neo4j, domain, computers, users):
         self.arguments = arguments
         self.neo4j = neo4j
+        self.domain = domain
+        self.computers = computers
+        self.users = users
         self.start = time.time()
         logger.print_debug("Computing other objects")
 
@@ -24,6 +30,8 @@ class Objects:
         self.dcsync_paths = neo4j.all_requests["set_dcsync1"]["result"] + neo4j.all_requests["set_dcsync2"]["result"]
 
         self.users_nb_domain_admins = neo4j.all_requests["nb_domain_admins"]["result"]
+
+        self.get_unpriv_users_to_GPO()
 
         end_nodes = []
         # Check if dcsync path is activated or not
@@ -178,3 +186,266 @@ class Objects:
         grid.setData(data)
         page.addComponent(grid)
         page.render()
+
+    
+    def get_unpriv_users_to_GPO(self):
+        if self.arguments.gpo_low and self.domain.unpriv_users_to_GPO is None:
+            return
+        if not self.arguments.gpo_low:
+            fail = []
+            if self.domain.unpriv_users_to_GPO_init is None:
+                fail.append("unpriv_users_to_GPO_init")
+            elif self.domain.unpriv_users_to_GPO_user_enforced is None:
+                fail.append("unpriv_users_to_GPO_user_enforced")
+            elif self.domain.unpriv_users_to_GPO_user_not_enforced is None:
+                fail.append("unpriv_users_to_GPO_user_not_enforced")
+            elif self.domain.unpriv_users_to_GPO_computer_enforced is None:
+                fail.append("unpriv_users_to_GPO_computer_enforced")
+            elif self.domain.unpriv_users_to_GPO_computer_not_enforced is None:
+                fail.append("unpriv_users_to_GPO_computer_not_enforced")
+
+            if 0 < len(fail) < 5:  # if only some of them are disabled
+                logger.print_error(
+                    f" In order to use 'normal GPO mode', please activate the following in config.json : {', '.join(fail)}"
+                )
+
+            if len(fail) > 0:
+                return
+
+        def parseGPOData(listOfPaths, headers):
+            """
+            Initial parsing of data from neo4j requests for GPO
+            """
+            dictOfGPO = {}
+            for path in listOfPaths:
+                start = path.nodes[0]
+                end = path.nodes[-1]
+                if "GPO" in start.labels:
+                    nameOfGPO = start.name
+                    idOfGPO = start.id
+                    sens = "right"
+                elif "GPO" in end.labels:
+                    nameOfGPO = end.name
+                    idOfGPO = end.id
+                    sens = "left"
+                else:
+                    continue
+                try:
+                    if sens == "right":
+                        dictOfGPO[nameOfGPO][headers[4]] += 1
+                        dictOfGPO[nameOfGPO]["right_path"].append(path)
+                        dictOfGPO[nameOfGPO]["end_list"].append(end.name)
+                    elif sens == "left":
+                        dictOfGPO[nameOfGPO][headers[1]] += 1
+                        dictOfGPO[nameOfGPO]["left_path"].append(path)
+                        dictOfGPO[nameOfGPO]["entry_list"].append(start.name)
+                    else:
+                        continue
+                except KeyError:
+                    if sens == "right":
+                        dictOfGPO[nameOfGPO] = {
+                            headers[0]: nameOfGPO,
+                            headers[1]: 0,
+                            headers[4]: 1,
+                            "left_path": [],
+                            "right_path": [path],
+                            "id": idOfGPO,
+                            "entry_list": [],
+                            "end_list": [end.name],
+                        }
+                    elif sens == "left":
+                        dictOfGPO[nameOfGPO] = {
+                            headers[0]: nameOfGPO,
+                            headers[1]: 1,
+                            headers[4]: 0,
+                            "left_path": [path],
+                            "right_path": [],
+                            "id": idOfGPO,
+                            "entry_list": [start.name],
+                            "end_list": [],
+                        }
+                    else:
+                        continue
+            return dictOfGPO
+
+        def formatGPOGrid(dictOfGPO, headers):
+            output = []
+
+            # Extract all computers admin of computers
+            computers_with_admin_rights = [d["Computer Admin"].split('</i> ')[-1] for d in self.computers.computers_admin_data_grid]
+            # Extract all users admin of computers
+            users_with_admin_rights = [d["User"].split('</i> ')[-1] for d in self.users.users_admin_of_computers]
+
+            for _, dict in dictOfGPO.items():
+                self.domain.number_of_gpo += 1
+                # Rate the interest of the GPO
+                # 0 star  : no object impacted
+                # 1 star  : at least one object impacted
+                # 2 stars : at least one admin account impacted
+                # 3 stars : full domain or at least one domain admin impacted
+                paths = dict["right_path"]
+
+                if len(paths) == 0:
+                    interest = 0
+                else:
+                    interest = 1
+                
+                for path in paths:
+                    for i in range(len(path.nodes)):
+                        if path.nodes[i].labels == "Domain":
+                            interest = 3
+                            break
+                        if path.nodes[i].name in self.domain.admin_list:
+                            interest = 3
+                            break
+                        if path.nodes[i].name in users_with_admin_rights or path.nodes[i].name in computers_with_admin_rights:
+                            interest = max(2, interest)
+                
+                
+                icon = f"<span class='{interest}'></span><i class='bi bi-star-fill'></i>"*interest + "<i class='bi bi-star'></i>"*(3-interest)
+
+                output.append(
+                    {
+                        headers[0]: dict[headers[0]],
+                        headers[1]: f'<i class="bi bi-shuffle {str(dict[headers[1]]).zfill(6)}"></i> ' + str(dict[headers[1]]),
+                        headers[2]: {
+                            "link": "users_GPO_access-%s-left-graph.html"
+                            % (quote(str(dict[headers[0]]).replace(sep, '_'))),
+                            "value": "<i class='bi bi-diagram-3-fill' aria-hidden='true'></i>",
+                        },
+                        headers[3]: {
+                            "link": "users_GPO_access-%s-left-grid.html"
+                            % (quote(str(dict[headers[0]]).replace(sep, '_'))),
+                            "value": "<i class='bi bi-list-columns-reverse' aria-hidden='true'></i>",
+                        },
+                        headers[4]: f'<i class="bi bi-bullseye {str(len(list(set(dict["end_list"])))).zfill(6)}"></i> ' + str(len(list(set(dict["end_list"])))),
+                        headers[5]: icon,
+                        headers[6]: {
+                            "link": "users_GPO_access-%s-right-graph.html"
+                            % (quote(str(dict[headers[0]]).replace(sep, '_'))),
+                            "value": "<i class='bi bi-diagram-3-fill' aria-hidden='true'></i>",
+                        },
+                        headers[7]: {
+                            "link": "users_GPO_access-%s-right-grid.html"
+                            % (quote(str(dict[headers[0]]).replace(sep, '_'))),
+                            "value": "<i class='bi bi-list-columns-reverse' aria-hidden='true'></i>",
+                        },
+                    }
+                )
+            return output
+
+        def formatSmallGrid(list, gpo_name):
+            output = []
+            for n in list:
+                output.append({gpo_name: n})
+            return output
+
+        headers = [
+            "GPO name",
+            "Paths to GPO",
+            "Inbound graph",
+            "Inbound list",
+            "Objects impacted",
+            "Targets interest",
+            "Outbound graph",
+            "Outbound list",
+        ]
+        if not self.arguments.gpo_low:
+            data = (
+                self.domain.unpriv_users_to_GPO_init
+                + self.domain.unpriv_users_to_GPO_user_enforced
+                + self.domain.unpriv_users_to_GPO_computer_enforced
+                + self.domain.unpriv_users_to_GPO_user_not_enforced
+                + self.domain.unpriv_users_to_GPO_computer_not_enforced
+            )
+            self.domain.unpriv_users_to_GPO_parsed = parseGPOData(data, headers)
+            grid = Grid("Users with GPO access")
+        else:
+            self.domain.unpriv_users_to_GPO_parsed = parseGPOData(
+                self.domain.unpriv_users_to_GPO, headers
+            )
+            grid = Grid("Users with GPO access")
+
+        formated_data = sorted(
+            formatGPOGrid(self.domain.unpriv_users_to_GPO_parsed, headers),
+            key=lambda x: x[headers[1]],
+            reverse=True,
+        )
+        page = Page(
+            self.domain.arguments.cache_prefix,
+            "users_GPO_access",
+            "Exploitation through GPO",
+            "users_GPO_access",
+        )
+
+        grid.setheaders(headers)
+        grid.setData(json.dumps(formated_data))
+        page.addComponent(grid)
+        page.render()
+
+        for _, GPO in self.domain.unpriv_users_to_GPO_parsed.items():
+            url_left_graph = "users_GPO_access-%s-left-graph" % GPO[headers[0]]
+            url_right_graph = "users_GPO_access-%s-right-graph" % GPO[headers[0]]
+            page_left_graph = Page(
+                self.arguments.cache_prefix,
+                url_left_graph,
+                "Users with write access on GPO",
+                "graph_GPO_access",
+            )
+            page_right_graph = Page(
+                self.arguments.cache_prefix,
+                url_right_graph,
+                "Objects impacted by GPO",
+                "graph_GPO_access",
+            )
+
+            url_left_grid = "users_GPO_access-%s-left-grid" % GPO[headers[0]]
+            url_right_grid = "users_GPO_access-%s-right-grid" % GPO[headers[0]]
+            page_left_grid = Page(
+                self.arguments.cache_prefix,
+                url_left_grid,
+                "List of users able to compromise %s" % GPO[headers[0]],
+                "grid_GPO_access",
+            )
+            page_right_grid = Page(
+                self.arguments.cache_prefix,
+                url_right_grid,
+                "List of users impacted by %s" % GPO[headers[0]],
+                "grid_GPO_access",
+            )
+
+            # if GPO[headers[4]] > 0:
+            graph_left = Graph()
+            graph_left.setPaths(GPO["left_path"])
+            page_left_graph.addComponent(graph_left)
+
+            graph_right = Graph()
+            graph_right.setPaths(GPO["right_path"])
+            page_right_graph.addComponent(graph_right)
+
+            if not self.arguments.gpo_low:
+                entry_grid = Grid("List of users able to compromise %s" % GPO[headers[0]])
+            else:
+                entry_grid = Grid("List of users able to compromise %s" % GPO[headers[0]])
+            entry_grid.setheaders([GPO[headers[0]]])
+            entry_grid.setData(
+                json.dumps(
+                    formatSmallGrid(
+                        list(set(GPO["entry_list"])), GPO[headers[0]])
+                )
+            )
+            page_left_grid.addComponent(entry_grid)
+
+            
+            end_grid = Grid("List of users impacted by %s" % GPO[headers[0]])
+            end_grid.setheaders([GPO[headers[0]]])
+            end_grid.setData(
+                json.dumps(formatSmallGrid(
+                    list(set(GPO["end_list"])), GPO[headers[0]]))
+            )
+            page_right_grid.addComponent(end_grid)
+
+            page_left_graph.render()
+            page_right_graph.render()
+            page_left_grid.render()
+            page_right_grid.render()
